stdio-mgr: do not propagate attrs
I did this originally to solve a problem in sphinxcadquery, but then I
started getting two versions of attrs in an output and sphinxcadquery
wouldn't build, so heres the patch to fix it. Sphinxcadquery probably
doesn't work now, but since CQ has replaced it with VTK I'm just leaving
sphinxcadquery here as an archive.